pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dh-converter"
version = "0.1.0"
description = "Convert 5e monsters to Daggerheart adversaries (rules-driven)."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Your Name"}]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "Jinja2>=3.1"
]

[project.scripts]
dh-convert = "converter.__main__:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = [""]
include = ["converter*"]
